+++ src/containerapp/HISTORY.rst
@@ -4,6 +4,9 @@ Release History
 ===============
 upcoming
 ++++++
+
+1.1.0b5
+++++++
 * 'az containerapp auth update': Support authenticating blob storage token store using managed identity with `--blob-container-uri` and `--blob-container-identity`.
 * 'az containerapp env create': Set identity only when `--mi-system-assigned` or `--mi-user-assigned` is specified.
 * 'az containerapp env create': Set identity only when `--system-assigned` or `--user-assigned` is specified.